backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
flask==3.1.1
flask-sock==0.7.0
pyserial==3.5
